+++ CHANGELOG.md
@@ -7,8 +7,8 @@
 - module `mod3rd.admin_esp.softap` for SoftAP configuration
 - `start_auto_config` supports different timezone handler with `modext.misc.boot.set_timezone`
 - auto-discovery-configuration (`ADC`) and `auto_config` loader sets the plugin `path_spec`
-- modcore `lifecycle hooks` for more flexible enhanced integration (also with asyncio)
- sample code in `boot.py`
+- modcore [`lifecycle hooks`](https://github.com/kr-g/mpymodcore/blob/master/modcore/lifecycle.py)
+ for more flexible enhanced integration (also with asyncio). sample code in `boot.py`
 -