+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 
 # Changelog
 
-## next version v0.0.18
+## version v0.0.18
 
 - auto-discovery-configuration `__app__.apt_ext` can be a single Plugin, or a list of Plugin
 - module `mod3rd.admin_esp.softap` for SoftAP configuration